add unique tiger indexes, and multicolumn indexes
add truncate! method

add county specific partitions, and the appropriate check constraint.
This is because there's no easy way with ogr2ogr to get rid of duplicates where a
feature is in more than one county (we could
create a temp table, then a select distinct, but... ewww, this is probably
better for queries anyway).